Former-commit-id: 857b654c277c59c8e8adc3df0678a769bb3a8998

+++ ifs-data-service/src/main/java/org/innovateuk/ifs/publiccontent/controller/PublicContentItemController.java
@@ -22,8 +22,8 @@ public class PublicContentItemController {
     @RequestMapping(value = "find-by-filter", method = RequestMethod.POST)
     public RestResult<PublicContentItemPageResource> findFilteredItems(@RequestParam(value = "innovationAreaId", required = false) Optional<Long> innovationAreaId,
                                                                        @RequestParam(value = "searchString", required = false) Optional<String> searchString,
-                                                                       @RequestParam(value = "pageNumber", required = false) Optional<Long> pageNumber,
-                                                                       @RequestParam(value = "pageSize", required = false) Optional<Long> pageSize) {
+                                                                       @RequestParam(value = "pageNumber", required = false) Optional<Integer> pageNumber,
+                                                                       @RequestParam(value = "pageSize", required = false) Optional<Integer> pageSize) {
         return publicContentItemService.findFilteredItems(innovationAreaId, searchString, pageNumber, pageSize).toGetResponse();
     }
